+++ src/components/Input/Input.module.css
@@ -7,7 +7,7 @@
 /**
  * Default text input styles 
  */
-:where(.input-field) {
+:where(.input) {
   @mixin inputStyles;
   padding-left: var(--eds-size-2);
 }
