+++ src/components/Input/Input.test.ts
@@ -1,0 +1,6 @@
+import { generateSnapshots } from '@chanzuckerberg/story-utils';
+import * as InputStoryFile from './Input.stories';
+
+describe('<Input />', () => {
+  generateSnapshots(InputStoryFile);
+});
