+++ src/components/Input/index.ts
@@ -1,0 +1,2 @@
+export { Input as default } from './Input';
+export type { InputProps } from './Input';
