--- src/components/InputField/InputField.test.ts
@@ -1,6 +1,0 @@
-import { generateSnapshots } from '@chanzuckerberg/story-utils';
-import * as InputFieldStoryFile from './InputField.stories';
-
-describe('<InputField />', () => {
-  generateSnapshots(InputFieldStoryFile);
-});
